+++ src/main/kotlin/data/repo/sql/DataBlocks.kt
@@ -4,7 +4,7 @@ import org.jetbrains.exposed.sql.Table
 
 object DataBlocks : Table("data_blocks") {
     val hash = text("hash").uniqueIndex()
-    val size = integer("size").check { it.greaterEq(0) }
+    val size = long("size").check { it.greaterEq(0) }
 
     override val primaryKey = PrimaryKey(hash)
 }
